requirements.txt
Flask
asyncio
aiohttp
pycryptodome
protobuf
gunicorn
